fix compile problem on FreeBSD
git-svn-id: svn://svn.code.sf.net/p/cxxtools/code/trunk@1563 15ae5fad-cc11-0410-8fac-bce609e504b0

+++ cxxtools/src/log.cpp
@@ -170,7 +170,7 @@ namespace cxxtools
       char* p = putInt(str, getpid());
       entry.append(str, p - str);
       entry += '.';
-      p = putInt(str, pthread_self());
+      p = putInt(str, (unsigned long)pthread_self());
       entry.append(str, p - str);
       entry += "] ";
       entry += level;
